fix: Prevent lengthy spinner on load when selected network is slow (#35516)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

When starting, the UI creates a StreamProvider that wraps the MetaMask
provider and can be used to make internal and external requests through
the core JSON-RPC pipeline via the background connection. This
StreamProvider must be initialized in order for the UI to complete the
start sequence and in order for users to see anything on screen.

When the StreamProvider initializes, it will make a background request
for `metamask_getProviderState` to populate its internal state. As a
part of its logic, this RPC method will retrieve the network version
from the dapp-selected RPC endpoint. However, this creates a problem. If
the RPC endpoint is slow, this request won't resolve in a timely manner,
and the user will continue to see a spinner.

To address this, this commit changes the `metamask_getProviderState`
handler to accept a parameter, `isInitializingStreamProvider`. It
expects that StreamProvider will call the RPC method and pass `true` for
this parameter (this happens in `@metamask/provider` 22.1.1). When this
happens, the request for `net_version` will not be made and the initial
network version will be set to "loading". This allows the UI to complete
its start sequence regardless of how slow the network is. Once the UI is
loaded, it will update the `memStore` state, and
`metamask_getProviderState` will be requested again, but this time,
`isInitializingStreamProvider` will be set to `false`. This means the
network version will be properly retrieved and propagated to all of the
places that need it.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/35516?quickstart=1)

## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the
CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
2. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing
description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and
accurately)
-->

CHANGELOG entry: Prevent lengthy spinner on load when selected network
is slow to respond

## **Related issues**

Related to https://github.com/MetaMask/metamask-extension/issues/34214

## **Manual testing steps**

- Check out the `main` branch.
- Find `getProviderNetworkState` in `metamask-controller.js` and add the
following:
    ``` diff
let networkVersion = this.deprecatedNetworkVersions[networkClientId];
          if (networkVersion === undefined && completedOnboarding) {
    +       await new Promise((resolve) => {
    +         setTimeout(() => resolve(), 15000);
    +       })
            try {
              const result = await networkClient.provider.request({
                method: 'net_version',
              });
              networkVersion = convertNetworkId(result);
            } catch (error) {
              console.error(error);
              networkVersion = null;
            }
      
this.deprecatedNetworkVersions[networkClientId] = networkVersion;
          }
    ```
- Run `yarn start`.
- Open the popup. You should see a spinner for 15 seconds before the
home screen is shown.
- Now check out this branch.
- Reapply the same patch as above.
- Wait for the extension to rebundle.
- Reload the extension.
- Open the popup. The spinner should briefly show but then you should
see the home screen much more quickly.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**


https://github.com/user-attachments/assets/b9cb3fc5-ac2d-4bff-a338-a2604b814e8a

### **After**


https://github.com/user-attachments/assets/2f3f441a-0af5-4d46-952a-d367d58c746d

## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.