fix: connect now modal for bip44 accounts. (#36064)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

This PR fixes the issue where the connect now modal is not working when
the selected account has not been connected to the dapp.

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/36064?quickstart=1)

## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the
CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
2. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing
description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and
accurately)
-->

CHANGELOG entry: Fix connect now modal to support bip44 accounts.

## **Related issues**

Fixes:
https://consensyssoftware.atlassian.net/browse/MUL-868?atlOrigin=eyJpIjoiOTI4NDdlMzU5NDk1NDMyNWE4OWJmNmJmZDBiYWExYTgiLCJwIjoiaiJ9

## **Manual testing steps**
Add this to `.manifest-overrides.json`
```
"enableMultichainAccounts": {
  "enabled": true,
  "featureVersion": "2",
  "minimumVersion": "12.19.0"
}
 ```

1. Go to https://metamask.github.io/test-dapp-multichain/latest/
2. Connect only a solana account
3. Open the wallet 
4. Switch to another account group with a solana account
5. See the connect now modal
6. Click on connect now, see that the account is now connected. 

## **Screenshots/Recordings**

### **Before**

https://github.com/user-attachments/assets/e0fd38c6-0ce3-420a-9065-ca331a614224

### **After**

https://github.com/user-attachments/assets/5d429416-6a09-4d0d-81d1-b7dfda7289ef

## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Extension Coding Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format if applicable
- [x] I’ve applied the right labels on the PR (see [labeling guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)). Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described in the ticket it closes and includes the necessary testing evidence such as recordings and or screenshots.