fix: Show transactions for non-EVM and non-popular networks on the activity list (#36288)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

The activity list was not showing transactions on the Solana network.
This is because in
https://github.com/MetaMask/metamask-extension/pull/35689/ the
multichain-transactions-controller version was updated with a breaking
change. The MultichainTransactionControllerState was updated to nest
TransactionStateEntry per chain as well as the preexisting accountId:

```
export type MultichainTransactionsControllerState = {
    nonEvmTransactions: {
        [accountId: string]: {
            [chain: CaipChainId]: TransactionStateEntry;
        };
    };
};
```

See app/scripts/migrations/176.ts for more details.

This means that on `getSelectedAccountGroupMultichainTransactions`,
transactions need to be checked knowing that nonEvmTransactionsByAccount
is keyed by accountId but also by chainIds. In
unified-transaction-list.component we pass nonEvmChainIds to this
selector

Additionally, the activity list was not showing transactions on the
sepolia network. The root cause is that getAllNetworkTransactions (that
is used by nonceSortedCompletedTransactionsSelectorAllChains ->
nonceSortedTransactionsSelectorAllChains ->
selectedAddressTxListSelectorAllChain) filters transactions by chainid
of popular networks.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/36288?quickstart=1)

## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the
CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
2. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing
description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and
accurately)
-->

CHANGELOG entry:

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/36205

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.